Moved functions to separate files (#7)
Because the docstrings can take lots of space, and example code/tests
can be quite verbose I took the decision to move each of the
functions to a separate file.
