+++ src/main.rs
@@ -5,7 +5,7 @@ mod parser;
 use clap::Parser;
 use simple_logger::SimpleLogger;
 
-use crate::parser::{find_commands, utils};
+use crate::parser::{find_commands, read_lines};
 
 fn main() {
     SimpleLogger::new().init().unwrap();
