+++ src/parser/mod.rs
@@ -1,4 +1,6 @@
 //! Contains all of the parser logic for finding the LaTeX commands.
 
+pub mod check_line;
+pub mod find_brackets;
 pub mod find_commands;
-pub mod utils;
+pub mod read_lines;
